+++ RecieptHub/DTO/DishIngredientDto.cs
@@ -6,21 +6,21 @@ public class DishIngredientDto
     public int DishId { get; set; }
     public int IngredientId { get; set; }
     public string? IngredientName { get; set; }
-    public decimal QuantityGrams { get; set; }
-    public decimal CaloriesInDish { get; set; }
-    public decimal ProteinsInDish { get; set; }
-    public decimal FatsInDish { get; set; }
-    public decimal CarbohydratesInDish { get; set; }
+    public float QuantityGrams { get; set; }
+    public float CaloriesInDish { get; set; }
+    public float ProteinsInDish { get; set; }
+    public float FatsInDish { get; set; }
+    public float CarbohydratesInDish { get; set; }
 }
 
 public class DishIngredientCreateDto
 {
     public int DishId { get; set; }
     public int IngredientId { get; set; }
-    public decimal QuantityGrams { get; set; }
+    public float QuantityGrams { get; set; }
 }
 
 public class DishIngredientUpdateDto
 {
-    public decimal QuantityGrams { get; set; }
+    public float QuantityGrams { get; set; }
 }
