+++ RecieptHub/DTO/IngredientDto.cs
@@ -13,10 +13,10 @@ public class IngredientDto
 public class IngredientCreateDto
 {
     public required string Name { get; set; }
-    public decimal Calories { get; set; }
-    public decimal Proteins { get; set; }
-    public decimal Fats { get; set; }
-    public decimal Carbohydrates { get; set; }
+    public float Calories { get; set; }
+    public float Proteins { get; set; }
+    public float Fats { get; set; }
+    public float Carbohydrates { get; set; }
 }
 
 public class IngredientUpdateDto
